--- model/contacts.json
@@ -1,1 +1,0 @@
-[{"id":1,"name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":2,"name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":3,"name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":4,"name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":5,"name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":6,"name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":7,"name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":8,"name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":9,"name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":10,"name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":11,"name":"Ilfatey","email":"ialiev@gmail.com","phone":"89509499961"},{"id":12,"name":"Alfred","email":"alfred@gmail.com","phone":"8950963354"},{"id":13,"name":"ArchibaldVII","email":"ArchibaldVII@gmail.com","phone":"89600563563"},{"id":14,"name":"ivan","email":"ivan@gmail.com","phone":"89600563563"},{"id":15,"name":"Ilfat","email":"ialiev@mail.ru","phone":"89509499961"},{"id":16,"name":"Jack","email":"Jack@gmail.com","phone":"89600555555"}]
