+++ model/contacts/normalizationData.js
@@ -1,0 +1,6 @@
+const normalizationData = (data) => {
+  const dataToSting = data.toString()
+  return JSON.parse(dataToSting)
+}
+
+module.exports = normalizationData
