+++ routes/api/contacts.js
@@ -1,101 +1,24 @@
 const express = require('express')
 const router = express.Router()
-const Joi = require('joi')
 
-const controlContacts = require('../../model/index')
+const {
+  createContactController,
+  getContactByIdController,
+  getListContactsController,
+  removeContactByIdController,
+  updateContactByIdController
+} = require('../../controllers/contacts')
 
-router.get('/', async (req, res, next) => {
-  const data = await controlContacts.listContacts()
-  res.status(200).json(data)
-})
+const { validationCreateContact, validationUpdateContact } = require('../../midlevares/validation')
 
-router.get('/:contactId', async (req, res, next) => {
-  const { contactId } = req.params
+router.get('/', getListContactsController)
 
-  const data = await controlContacts.getContactById(contactId)
-  data ? res.status(200).json(data) : res.status(404).json({ message: 'Not found' })
-})
+router.get('/:contactId', getContactByIdController)
 
-router.post('/', async (req, res, next) => {
-  const shema = Joi.object({
-    name: Joi.string()
-      .alphanum()
-      .min(2)
-      .max(10)
-      .required(),
-    email: Joi.string()
-      .email()
-      .required(),
-    phone: Joi.number()
-      .integer()
-      .min(89000000000)
-      .max(89999999999)
-      .required()
-  })
+router.post('/', validationCreateContact, createContactController)
 
-  const { error } = shema.validate(req.body)
+router.delete('/:contactId', removeContactByIdController)
 
-  if (error) {
-    const text = error?.details[0].message.replace(/["]/g, '')
-    res.status(400).json({ message: text })
-    return
-  }
-
-  const updateData = await controlContacts.addContact(req.body)
-  res.status(201).json(updateData)
-})
-
-router.delete('/:contactId', async (req, res, next) => {
-  const { contactId } = req.params
-
-  const data = await controlContacts.removeContact(contactId)
-  data.status === 200
-    ? res.status(200).json({ message: 'contact deleted' })
-    : res.status(400).json({ message: 'Not found' })
-})
-
-router.patch('/:contactId', async (req, res, next) => {
-  const { contactId } = req.params
-
-  const shema = Joi.object({
-    name: Joi.string()
-      .alphanum()
-      .min(2)
-      .max(10)
-      .allow('')
-      .required(),
-    email: Joi.string()
-      .email()
-      .allow('')
-      .required(),
-    phone: Joi.number()
-      .integer()
-      .min(89000000000)
-      .max(89999999999)
-      .allow('')
-      .required(),
-
-  })
-
-  const { error, value } = shema.validate(req.body)
-  console.log(value)
-
-  if (error) {
-    const text = error?.details[0].message.replace(/["]/g, '')
-    res.status(400).json({ message: text })
-    return
-  }
-  const obj = Object.keys(value)
-  const updateItems = {}
-
-  for (const key of obj) {
-    if (req.body[key].trim().length !== 0) updateItems[key] = req.body[key]
-  }
-
-  const data = await controlContacts.updateContact(contactId, updateItems)
-  data
-    ? res.status(200).json(data)
-    : res.status(404).json({ message: 'Not found' })
-})
+router.patch('/:contactId', validationUpdateContact, updateContactByIdController)
 
 module.exports = router
